# lab4/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  serializer.CurrentJSON:
    properties:
      created_at:
        type: string
      creator_login:
        type: string
      current_id:
        type: integer
      finish_date:
        type: string
      form_date:
        type: string
      moderator_login:
        type: string
      status:
        type: string
      voltage_bord:
        type: number
    type: object
...
  serializer.StatusJSON:
    properties:
      status:
        description: Статус заявки
        type: string
    type: object
...
host: localhost
info:
  contact:
    email: support@current.com
    name: API Support
    url: http://localhost
  description: API для определения необходимой силы тока
  license:
    name: MIT
  title: Current API
  version: "1.0"
paths:
  /current-calculations/{id}:
    get:
      description: Возвращает полную информацию о заявке
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Данные заявки с устройствами
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Неверный ID
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещен
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Заявка не найдено
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Получить заявку по ID
      tags:
      - Currents
  /current-calculations/{id}/delete-current-calculations:
    delete:
      description: Выполняет логическое удаление заявки
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Статус удаления
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Неверный запрос
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещен
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Заявка не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Удалить заявка
      tags:
      - Currents
  /current-calculations/{id}/edit-current-calculations:
    put:
      consumes:
      - application/json
      description: Обновляет данные заявки
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      - description: Новые данные заявки
        in: body
        name: current
        required: true
        schema:
          $ref: '#/definitions/serializer.CurrentJSON'
      produces:
      - application/json
      responses:
        "200":
          description: Обновленная заявка
          schema:
            $ref: '#/definitions/serializer.CurrentJSON'
        "400":
          description: Неверные данные
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Заявка не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Изменить заявка
      tags:
      - Currents
  /current-calculations/{id}/finish:
    put:
      consumes:
      - application/json
      description: Изменяет статус заявки (только для модераторов)
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      - description: Новый статус
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/serializer.StatusJSON'
      produces:
      - application/json
      responses:
        "200":
          description: Результат модерации
          schema:
            $ref: '#/definitions/serializer.CurrentJSON'
        "400":
          description: Неверный запрос
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещен
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Заявка не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Завершить заявку
      tags:
      - Currents
  /current-calculations/{id}/form:
    put:
      description: Переводит заявку в статус "formed"
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Сформированная заявка
          schema:
            $ref: '#/definitions/serializer.CurrentJSON'
        "400":
          description: Неверный запрос
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещен
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Заявка не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Сформировать заявку
      tags:
      - Currents
  /current-calculations/current-calculations:
    get:
      description: Возвращает заявки с возможностью фильтрации по датам и статусу
      parameters:
      - description: Начальная дата (YYYY-MM-DD)
        in: query
        name: from-date
        type: string
      - description: Конечная дата (YYYY-MM-DD)
        in: query
        name: to-date
        type: string
      - description: Статус заявки
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список заявок
          schema:
            items:
              $ref: '#/definitions/serializer.CurrentJSON'
            type: array
        "400":
          description: Неверный формат даты
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Получить список заявок на расчёт
      tags:
      - Currents
  /current-calculations/current-cart:
    get:
      description: Возвращает информацию о текущей заявке-черновике на расчёт пользователя
      produces:
      - application/json
      responses:
        "200":
          description: Данные корзины заявки-черновика
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Неверный запрос
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Получить корзину расчёта
      tags:
      - Currents
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"